+++ hermes/bin/src/runtime_extensions/hermes/binary/host.rs
@@ -1,0 +1,5 @@
+//! Binary host implementation for WASM runtime.
+
+use crate::{runtime_extensions::bindings::hermes::binary::api::Host, state::HermesState};
+
+impl Host for HermesState {}
