+++ hermes/bin/src/runtime_extensions/hermes/cbor/host.rs
@@ -1,0 +1,5 @@
+//! CBOR host implementation for WASM runtime.
+
+use crate::{runtime_extensions::bindings::hermes::cbor::api::Host, state::HermesState};
+
+impl Host for HermesState {}
