+++ hermes/bin/src/runtime_extensions/hermes/hash/mod.rs
@@ -1,15 +1,8 @@
-//! Host - Hash implementations
+//! Hash runtime extension implementation.
 
-use crate::{
-    runtime_extensions::{
-        bindings::hermes::{
-            binary::api::Bstr,
-            hash::api::{Errno, Host},
-        },
-        state::{Context, Stateful},
-    },
-    state::HermesState,
-};
+use crate::runtime_extensions::state::{Context, Stateful};
+
+mod host;
 
 /// State
 pub(crate) struct State {}
@@ -17,28 +10,5 @@ pub(crate) struct State {}
 impl Stateful for State {
     fn new(_ctx: &Context) -> Self {
         State {}
-    }
-}
-
-impl Host for HermesState {
-    /// Hash a binary buffer with BLAKE2s
-    fn blake2s(
-        &mut self, _buf: Bstr, _outlen: Option<u8>, _key: Option<Bstr>,
-    ) -> wasmtime::Result<Result<Bstr, Errno>> {
-        todo!()
-    }
-
-    /// Hash a binary buffer with `BLAKE2b`
-    fn blake2b(
-        &mut self, _buf: Bstr, _outlen: Option<u8>, _key: Option<Bstr>,
-    ) -> wasmtime::Result<Result<Bstr, Errno>> {
-        todo!()
-    }
-
-    /// Hash a binary buffer with BLAKE3
-    fn blake3(
-        &mut self, _buf: Bstr, _outlen: Option<u8>, _key: Option<Bstr>,
-    ) -> wasmtime::Result<Result<Bstr, Errno>> {
-        todo!()
     }
 }
