+++ hermes/bin/src/runtime_extensions/hermes/init/mod.rs
@@ -1,4 +1,4 @@
-//! Host - Init implementations
+//! Init runtime extension implementation.
 
 use crate::runtime_extensions::state::{Context, Stateful};
