+++ hermes/bin/src/runtime_extensions/hermes/json/host.rs
@@ -1,0 +1,5 @@
+//! JSON host implementation for WASM runtime.
+
+use crate::{runtime_extensions::bindings::hermes::json::api::Host, state::HermesState};
+
+impl Host for HermesState {}
