+++ hermes/bin/src/runtime_extensions/hermes/json/mod.rs
@@ -1,12 +1,8 @@
-//! Host - JSON implementations
+//! JSON runtime extension implementation.
 
-use crate::{
-    runtime_extensions::{
-        bindings::hermes::json::api::Host,
-        state::{Context, Stateful},
-    },
-    state::HermesState,
-};
+use crate::runtime_extensions::state::{Context, Stateful};
+
+mod host;
 
 /// State
 pub(crate) struct State {}
@@ -16,5 +12,3 @@ impl Stateful for State {
         State {}
     }
 }
-
-impl Host for HermesState {}
